requirements.txt
agno>=2.2.5
groq>=0.33.0
duckduckgo-search>=8.1.1
yfinance>=0.2.66
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas-ta>=0.4.71b0
mplfinance>=0.12.10b0
matplotlib>=3.7.0
numpy>=1.22.0
ddgs>=0.1.0
tabulate>=0.9.0

# Additional dependencies for enhanced features
scipy>=1.7.0
optuna>=2.10.0
quantstats>=0.0.37
backtesting>=0.3.3
scikit-learn>=1.0.0
python-dateutil>=2.8.2
pytz>=2021.3
plotly>=5.0.0
cufflinks>=0.17.3